rename: Program 관련 DTO 패키지 변경 #24
