+++ src/main/java/com/efub/dhs/domain/program/service/ProgramService.java
@@ -15,11 +15,11 @@ import com.efub.dhs.domain.program.dto.GoalDto;
 import com.efub.dhs.domain.program.dto.HostDto;
 import com.efub.dhs.domain.program.dto.ImageDto;
 import com.efub.dhs.domain.program.dto.NoticeDto;
-import com.efub.dhs.domain.program.dto.ProgramDetailResponseDto;
 import com.efub.dhs.domain.program.dto.ProgramDto;
 import com.efub.dhs.domain.program.dto.ProgramMemberDto;
-import com.efub.dhs.domain.program.dto.ProgramOutlineResponseDto;
-import com.efub.dhs.domain.program.dto.ProgramRegistryRequestDto;
+import com.efub.dhs.domain.program.dto.request.ProgramRegistryRequestDto;
+import com.efub.dhs.domain.program.dto.response.ProgramDetailResponseDto;
+import com.efub.dhs.domain.program.dto.response.ProgramOutlineResponseDto;
 import com.efub.dhs.domain.program.entity.Notice;
 import com.efub.dhs.domain.program.entity.Program;
 import com.efub.dhs.domain.program.entity.ProgramImage;
